made the rectangle a beyblade
So firstly i set the initial angle and the rotation speed. Then i add a clock for consistent rotation. Now i keep updating the rotation angle and create a rotation matrix. Then i pass the rectangle and its angle as the parameters in my rotation function and i maintain it to 144fps