Completed perspective projection
Projection of a 3D object on a 2D surface with depth perception meaning the object reduces in size when it moves away from us and increases in size when it comes closer to us.

1. Initialization
Here we initialize all of these values

2. Event
Here on pressing w, the object moves deeper into the z axis by 10px and on pressing s it moves outwards from the z axis by -10px

3. Loop
The angle here increases according to the rotation speed and stays within the 0 to 360 range

4. Rotation
We rotate each of the points, hence rotating the figure.

First we convert the angle to radians and then we use the mathematical formulas/matrices for x, y and z axis rotations

We rotate the 3D points across the Z axis then the Y axis and finally the X axis

5. Render
Here we plot the cube by plotting each and every point of the cube and then rotate the cube and project it and give the faces color

6. Project
The if condition is to ensure that the point isnt behind the camera

FOV
Field of view is basically the trapezium in which we can see. Our fov is a trapezium and not a cuboid as if it was a cuboid then we wouldnt know if an object is closer or away from us as it would always be the same size whereas in a trapezium, if an object moves away then its size decreases and if it comes close, its size increases which is the way our vision works.

Scaling Factor
This decides how much our object increases or decreases by when moving towards or away from us in the trapezium view.

Relation
Scaling factor = Fov / (Z + Distance)